chore: move transparentcanvas and streamingtext canvas under their own folder in components
