+++ src/components/Canvas/StreamingTextCanvas.jsx
