+++ src/components/Canvas/TransparentCanvas.jsx
